Add correct links for href for github pages
Finding out that the the redirections work a bit different was
interesting. The fact that using "/" or "/index.html" to go back to the
index page does not work in pages is not my favorite. Even with this
though the learning and the oppurtunity to find out exactly how it
works will make it much easier for the future.